reimplement gc-related metrics to match otel implementation

+++ tracer/src/Datadog.Trace/RuntimeMetrics/PerformanceCounterWrapper.cs
@@ -33,7 +33,7 @@ namespace Datadog.Trace.RuntimeMetrics
             _counter?.Dispose();
         }
 
-        public double? GetValue(string instanceName)
+        public long? GetValue(string instanceName)
         {
             var counter = _counter;
